// packages/components/src/ui/Breadcrumbs/Breadcrumbs.styles.d.ts
export declare const useStyles: () => {
  container: {
    fontFamily:
      | 'Barlow-Light'
      | 'Barlow-Regular'
      | 'Barlow-Medium'
      | 'Barlow-SemiBold'
      | 'Barlow-Bold'
      | 'Inter-Light'
      | 'Inter-Regular'
      | 'Inter-Medium'
      | 'Inter-SemiBold'
      | 'Inter-Bold';
    fontSize: 12;
    lineHeight: 20;
    letterSpacing: 0;
    flexDirection: 'row';
    flexWrap: 'wrap';
    alignItems: 'center';
    gap: import('@alveole/theme').Spacing;
  };
  separator: {
    color: '#5F6571';
  };
  current: {
    color: '#151617';
  };
  link: {
    transitionProperty: 'text-decoration';
    transitionDuration: '0.1s';
    transitionTimingFunction: 'ease-in-out';
    fontFamily:
      | 'Barlow-Light'
      | 'Barlow-Regular'
      | 'Barlow-Medium'
      | 'Barlow-SemiBold'
      | 'Barlow-Bold'
      | 'Inter-Light'
      | 'Inter-Regular'
      | 'Inter-Medium'
      | 'Inter-SemiBold'
      | 'Inter-Bold';
    fontSize: 12;
    lineHeight: 20;
    letterSpacing: 0;
    color: '#5F6571';
    textDecoration: string;
  };
  linkHover: {
    textDecoration: string;
  };
};
//# sourceMappingURL=Breadcrumbs.styles.d.ts.map